Move Verse Node
VerseNode isn't really a primitive to be used of all of Orature, but just a data type used in the Narration package.
